quotations-streamlit/requirements.txt
google-generativeai==0.8.5
pandas==2.2.3
openpyxl==3.1.5
xlrd== 2.0.1
python-dotenv==1.1.0
slack_sdk==3.35.0
streamlit==1.31.0
numpy==1.26.4
matplotlib==3.7.2